requirements.txt
aenum==3.1.11
alembic==1.8.0
anyio==3.6.1
APScheduler==3.9.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==22.2.0
azure-core==1.24.2
azure-cosmos==4.3.0
azure-storage-blob==12.13.0
bcrypt==3.2.2
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
Deprecated==1.2.13
Django==4.0.4
django-extensions==3.1.5
ecdsa==0.18.0
exceptiongroup==1.1.1
fastapi==0.78.0
firebase-admin==5.2.0
google-api-core==2.8.2
google-api-python-client==2.54.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.6.0
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
greenlet==1.1.2
grpcio==1.47.0
grpcio-status==1.47.0
h11==0.13.0
httplib2==0.20.4
httptools==0.4.0
idna==3.3
iniconfig==2.0.0
isodate==0.6.1
Jinja2==3.1.2
kubernetes==24.2.0
Mako==1.2.1
MarkupSafe==2.1.1
msgpack==1.0.4
msrest==0.7.1
mysql-connector-python==8.0.29
mysqlclient==2.1.1
numpy==1.23.1
oauthlib==3.2.0
packaging==21.3
pandas==1.4.3
passlib==1.7.4
Pillow==9.2.0
pluggy==1.0.0
proto-plus==1.20.6
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.9.1
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.2.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==4.3.4
reportlab==3.6.11
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
sendgrid==6.9.7
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.10
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.15.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
websockets==10.3
wrapt==1.14.1
XlsxWriter==3.0.3


